+++ tests/Sqlsrv/UpdateTest.php
@@ -7,6 +7,11 @@ class UpdateTest extends Common\UpdateTest
 {
     protected $db_type = 'sqlsrv';
 
+    protected function withRecursiveKeyword()
+    {
+        return 'WITH';
+    }
+
     /**
      * SQL Server has no IGNORE; asking for it has to say so rather than
      * fatal on an undefined method.